requirements.txt
glcontext==3.0.0
moderngl==5.12.0
numpy==2.2.2
pygame-ce==2.5.2
